requirements.txt
feedparser>=6.0.0
httpx>=0.25.0
anthropic>=0.18.0
pyyaml>=6.0
python-dateutil>=2.8.0
notion-client>=2.0.0
python-dotenv>=1.0.0

# Gmail API
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.100.0

# YouTube Transcript
youtube-transcript-api>=0.6.0

# Scheduler
schedule>=1.2.0

# Web Scraping (GitHub Trending)
beautifulsoup4>=4.12.0

# Web Application
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
jinja2>=3.1.0
python-multipart>=0.0.6
